#1 Cambio del tipo de la variable local thisAmount de la función getCharge() de la clase Rental, de int a double.

+++ src/ubu/gii/dass/refactoring/Rental.java
@@ -28,7 +28,7 @@ public class Rental {
 	}
 
 	public double getCharge() {
-		int thisAmount = 0;
+		double thisAmount = 0;
 		switch (getMovie().getPriceCode()) {
 		case Movie.REGULAR:
 			thisAmount += 2;
